extract route namespaces
