+++ src/lib/routes/about.ts
@@ -1,0 +1,7 @@
+import node from './primatives/node';
+
+const about = (r: string) => ({
+	id: (pageId: number) => node(`${r}/${pageId}`),
+});
+
+export default about;
