+++ src/lib/routes/bibles.ts
@@ -1,0 +1,13 @@
+import { Scalars } from '../generated/graphql';
+import node from './primatives/node';
+
+const bibles = (r: string) => ({
+	versionId: (versionId: Scalars['ID']) => node(`${r}/${versionId}`),
+	bookId: (bookId: Scalars['ID']) =>
+		node(`${r}/${bookId}`, (r) => ({
+			chapterNumber: (chapterNumber: Scalars['ID']) =>
+				node(`${r}/${chapterNumber}`),
+		})),
+});
+
+export default bibles;
