+++ src/lib/routes/discover.ts
@@ -1,0 +1,7 @@
+import node from './primatives/node';
+
+const discover = (r: string) => ({
+	collections: node(`${r}/collections`),
+});
+
+export default discover;
